Updated example with new combat logger

+++ character/stat.go
@@ -5,24 +5,24 @@ import (
 )
 
 type Stat struct {
-	value float32
+	value float64
 }
 
-func NewStat(value float32) *Stat {
+func NewStat(value float64) *Stat {
 	return &Stat{value: value}
 }
 
 type Modifier interface {
-	To(value float32)
-	Up(value float32)
-	Down(value float32)
+	To(value float64)
+	Up(value float64)
+	Down(value float64)
 }
 
 func (s *Stat) String() string {
 	return fmt.Sprintf("%v", s.value)
 }
 
-func (s *Stat) Up(value float32) {
+func (s *Stat) Up(value float64) {
 	if s.value+value <= 100.00 {
 		s.value += value
 		return
@@ -31,13 +31,13 @@ func (s *Stat) Up(value float32) {
 	s.value = 100.00
 }
 
-func (s *Stat) To(value float32) {
+func (s *Stat) To(value float64) {
 	if value <= 100.00 || value >= 0.00 {
 		s.value = value
 	}
 }
 
-func (s *Stat) Down(value float32) {
+func (s *Stat) Down(value float64) {
 	if s.value-value >= 0.00 {
 		s.value -= value
 		return
